+++ app/services/transform/functions/general.py
@@ -27,7 +27,7 @@ def any_exists(value: Value, values: list[Value], on_true: str = "1", on_false: 
     return on_false
 
 
-def lookup(value: Value, on_no_match: str = Empty, **kwargs: str) -> Value:
+def lookup(value: Value, on_no_match: str | Empty = Empty, **kwargs: str) -> Value:
     if value in kwargs:
         return kwargs[value]
     return on_no_match
