+++ app/transformers/flat.py
@@ -1,7 +1,7 @@
 from app.definitions.input import SurveyMetadata
 from app.definitions.spec import BuildSpec
 from app.services.formatters.formatter import Formatter
-from app.transformers.standard import SpecTransformer
+from app.transformers.spec import SpecTransformer
 
 
 class FlatSpecTransformer(SpecTransformer[SurveyMetadata, Formatter]):
