+++ app/transformers/looped.py
@@ -4,7 +4,7 @@ from app.definitions.input import SurveyMetadata
 from app.definitions.spec import BuildSpec, ParseTree
 from app.services.formatters.looping_formatter import LoopingFormatter
 from app.services.transform.populate import resolve_value_fields
-from app.transformers.standard import SpecTransformer
+from app.transformers.spec import SpecTransformer
 
 template_type = Literal["template", "looped"]
